Add custom scenarios widget to iD
The bulk of the work is done in a javascript file in the 'openmod.sh'
codebase. The only thing we have to do here is include the script
'custom_widgets.js' script in 'index.html' (which should be moved to the
openmod.sh repository anyway) and call the appropriate function to add
the widgets wherever we want them.

Note: it's important to load 'custom_widgets.js' prior to loading
      'iD.js'.