Update callback URI
This one's for the same reason as the previous commit, with the special
added feature, that we can finally have a relative, i.e. host/domain
independent callback URI.
This is really important as this means that we don't have to always
update the known callback URIs in `openmod.sh`.